[#176] Test infra modifications to run tests against yugabyted (#177)
Refactoring of test classes to improve dev-cycles for local testing. With this, the tests which are run against the container for YugabyteDB extend the class `YugabyteDBContainerTestBase` and to make them run locally with `yugabyted`, we will just need to extend the class `YugabytedTestBase` instead of the existing one.

There is one prerequisite for `yugabyted` i.e. it should run on `127.0.0.1`