Use container IP address to connect to YugabyteDB (#308)
## Problem

With the commit https://github.com/yugabyte/yugabyte-db/commit/cade0a8688fea889f1595bfa0a15d61124849f0d we cannot use the IP `0.0.0.0` to start the `yugabyted` process as it will start a process on `127.0.0.1` of the Docker container and thus we will not be able to connect it from outside the container i.e. the test framework.

This blocks the execution of all the tests since the YugabyteDB container itself will not be starting up with this.

## Solution

This PR implements a solution by leveraging the container IP address, so any RPC calls will be made to the container IP address.