Fixed test infra issue after PG15 merge (#370)
After the PG15 merge, upon running the tests, we are getting the following error in tests:

```
org.postgresql.util.PSQLException: ERROR: DROP DATABASE cannot be executed within a pipeline
```

This behaviour was originally introduced in PG14 and YugabyteDB inherited it upon its move to PG15. As a result, the tests which executed the DDL files by combining all the statements in a batch - they started failing.

To avoid the failure, this PR modifies the code so that the statements in the DDL files are now executed one at a time.